Fix build failed at first time.

+++ script/build.sh
@@ -11,7 +11,9 @@ debug() {
 }
 
 clean() {
-  rm ./build/CMakeCache.txt
+  if [ -f ./build/CMakeCache.txt ]; then
+    rm ./build/CMakeCache.txt
+  fi
 }
 
 build() {
